requirements.txt
google-adk
litellm
pydantic
numpy
google-cloud-aiplatform[adk,agent_engines]>=1.111
cloudpickle
